client/requirements.txt
streamlit>=1.25.0
mcp>=0.6.0
python-dotenv>=1.0.0
pyperclip>=1.8.2
fastapi>=0.95.0
uvicorn>=0.21.0
mcp-proxy>=0.5.1